Integrate visualization into cell_division example
Integrate `catalyst_adaptor` into scheduler
